+++ cpp/test/CMakeLists.txt
@@ -42,8 +42,6 @@ add_executable(
   common/sort.cpp
   fem/functionspace.cpp
   mesh/distributed_mesh.cpp
-  mesh/interval.cpp
-  mesh/rectangle.cpp
   mesh/generation.cpp
   common/CIFailure.cpp
   refinement/interval.cpp
